requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autobahn
Automat==22.10.0
backcall==0.2.0
black==23.7.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.6
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
decorator==5.1.1
Django==4.2.4
djangorestframework==3.14.0
executing==1.2.0
gunicorn==21.2.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.13
ipython
jedi==0.19.0
matplotlib-inline==0.1.6
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
priority==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
wcwidth==0.2.6
websockets==11.0.3
zope.interface==6.0
